Bring some consistency to PRs and Issues views - avatars in the issues view - rich hover in the PRs view - remove number from issue items
