+++ package.json
@@ -2561,11 +2561,6 @@
 					"when": "view == pr:github && viewItem == query"
 				},
 				{
-					"command": "issue.openIssue",
-					"when": "view == issues:github && viewItem =~ /^(link)?(current|continue)?issue/",
-					"group": "inline@1"
-				},
-				{
 					"command": "issue.chatSummarizeIssue",
 					"when": "view == issues:github && viewItem =~ /^(link)?(current|continue)?issue/ && github.copilot-chat.activated && config.githubPullRequests.experimental.chat",
 					"group": "inline@0"
