+++ src/issues/issuesView.ts
@@ -6,13 +6,14 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { commands, contexts } from '../common/executeCommands';
+import { issueMarkdown } from '../common/markdownUtils';
+import { DataUri } from '../common/uri';
 import { groupBy } from '../common/utils';
 import { FolderRepositoryManager, ReposManagerState } from '../github/folderRepositoryManager';
 import { IssueModel } from '../github/issueModel';
 import { RepositoriesManager } from '../github/repositoriesManager';
 import { issueBodyHasLink } from './issueLinkLookup';
 import { IssueItem, QueryGroup, StateManager } from './stateManager';
-import { issueMarkdown } from './util';
 
 export class QueryNode {
 	constructor(
@@ -74,11 +75,12 @@ export class IssuesTreeData
 		return new vscode.TreeItem(element.group, getQueryExpandState(this.context, element, element.isInFirstQuery ? vscode.TreeItemCollapsibleState.Expanded : vscode.TreeItemCollapsibleState.Collapsed));
 	}
 
-	private getIssueTreeItem(element: IssueItem): vscode.TreeItem {
-		const treeItem = new vscode.TreeItem(`${element.number}: ${element.title}`, vscode.TreeItemCollapsibleState.None);
-		treeItem.iconPath = element.isOpen
-			? new vscode.ThemeIcon('issues', new vscode.ThemeColor('issues.open'))
-			: new vscode.ThemeIcon('issue-closed', new vscode.ThemeColor('issues.closed'));
+	private async getIssueTreeItem(element: IssueItem): Promise<vscode.TreeItem> {
+		const treeItem = new vscode.TreeItem(element.title, vscode.TreeItemCollapsibleState.None);
+		treeItem.iconPath = (await DataUri.avatarCirclesAsImageDataUris(this.context, [element.author], 16, 16))[0] ??
+			(element.isOpen
+				? new vscode.ThemeIcon('issues', new vscode.ThemeColor('issues.open'))
+				: new vscode.ThemeIcon('issue-closed', new vscode.ThemeColor('issues.closed')));
 
 		treeItem.command = {
 			command: 'issue.openDescription',
@@ -103,7 +105,7 @@ export class IssuesTreeData
 		return treeItem;
 	}
 
-	getTreeItem(element: FolderRepositoryManager | QueryNode | IssueGroupNode | IssueItem): vscode.TreeItem {
+	async getTreeItem(element: FolderRepositoryManager | QueryNode | IssueGroupNode | IssueItem): Promise<vscode.TreeItem> {
 		if (element instanceof FolderRepositoryManager) {
 			return this.getFolderRepoItem(element);
 		} else if (element instanceof QueryNode) {
