+++ src/view/treeNodes/pullRequestNode.ts
@@ -270,7 +270,6 @@ export class PRNode extends TreeNode implements vscode.CommentingRangeProvider2 
 
 		const formattedPRNumber = number.toString();
 		let labelPrefix = currentBranchIsForThisPR ? '✓ ' : '';
-		let tooltipPrefix = currentBranchIsForThisPR ? 'Current Branch * ' : '';
 
 		if (
 			vscode.workspace
@@ -278,11 +277,9 @@ export class PRNode extends TreeNode implements vscode.CommentingRangeProvider2 
 				.get<boolean>(SHOW_PULL_REQUEST_NUMBER_IN_TREE, false)
 		) {
 			labelPrefix += `#${formattedPRNumber}: `;
-			tooltipPrefix += `#${formattedPRNumber}: `;
 		}
 
 		const label = `${labelPrefix}${isDraft ? '[DRAFT] ' : ''}${labelTitle}`;
-		const tooltip = `${tooltipPrefix}${title} by @${login}`;
 		const description = `by @${login}`;
 		const command = {
 			title: vscode.l10n.t('View Pull Request Description'),
@@ -293,7 +290,6 @@ export class PRNode extends TreeNode implements vscode.CommentingRangeProvider2 
 		return {
 			label,
 			id: `${this.parent instanceof TreeNode ? (this.parent.id ?? this.parent.label) : ''}${html_url}${this._isLocal ? this.pullRequestModel.localBranchName : ''}`, // unique id stable across checkout status
-			tooltip,
 			description,
 			collapsibleState: 1,
 			contextValue:
